[Upd] use patch + moved constants
